moved GetField and PutField methods from pkg val to sqle/index
