+++ go/libraries/doltcore/sqle/index/prolly_row_iter.go
@@ -22,12 +22,11 @@ import (
 	"github.com/dolthub/go-mysql-server/sql"
 
 	"github.com/dolthub/dolt/go/libraries/doltcore/schema"
-	"github.com/dolthub/dolt/go/store/pool"
 	"github.com/dolthub/dolt/go/store/prolly"
 	"github.com/dolthub/dolt/go/store/val"
 )
 
-type sqlRowIter struct {
+type prollyRowIter struct {
 	ctx  context.Context
 	iter prolly.MapRangeIter
 
@@ -38,7 +37,7 @@ type sqlRowIter struct {
 	rowLen  int
 }
 
-var _ sql.RowIter = sqlRowIter{}
+var _ sql.RowIter = prollyRowIter{}
 
 func NewProllyRowIter(ctx context.Context, sch schema.Schema, rows prolly.Map, rng prolly.Range, projections []string) (sql.RowIter, error) {
 	if schema.IsKeyless(sch) {
@@ -71,7 +70,7 @@ func rowIterFromMapIter(
 
 	kd, vd := m.Descriptors()
 
-	return sqlRowIter{
+	return prollyRowIter{
 		ctx:     ctx,
 		iter:    iter,
 		keyDesc: kd,
@@ -110,7 +109,7 @@ func projectionMappings(sch schema.Schema, projs []string) (keyMap, valMap val.O
 	return
 }
 
-func (it sqlRowIter) Next(ctx *sql.Context) (sql.Row, error) {
+func (it prollyRowIter) Next(ctx *sql.Context) (sql.Row, error) {
 	key, value, err := it.iter.Next(it.ctx)
 	if err != nil {
 		return nil, err
@@ -122,20 +121,18 @@ func (it sqlRowIter) Next(ctx *sql.Context) (sql.Row, error) {
 		if rowIdx == -1 {
 			continue
 		}
-		row[rowIdx] = it.keyDesc.GetField(keyIdx, key)
+		row[rowIdx] = GetField(it.keyDesc, keyIdx, key)
 	}
 	for valIdx, rowIdx := range it.valProj {
 		if rowIdx == -1 {
 			continue
 		}
-		row[rowIdx] = it.valDesc.GetField(valIdx, value)
+		row[rowIdx] = GetField(it.valDesc, valIdx, value)
 	}
 
 	return row, nil
 }
 
-func (it sqlRowIter) Close(ctx *sql.Context) error {
+func (it prollyRowIter) Close(ctx *sql.Context) error {
 	return nil
 }
-
-var shimPool = pool.NewBuffPool()
