+++ go/store/val/tuple_builder.go
@@ -16,11 +16,7 @@ package val
 
 import (
 	"encoding/json"
-	"fmt"
 	"time"
-
-	"github.com/dolthub/go-mysql-server/sql"
-	"github.com/dolthub/go-mysql-server/sql/expression/function"
 
 	"github.com/dolthub/dolt/go/store/pool"
 )
@@ -248,204 +244,4 @@ func (tb *TupleBuilder) PutRaw(i int, buf []byte) {
 	tb.fields[i] = tb.buf[tb.pos : tb.pos+sz]
 	writeBytes(tb.fields[i], buf)
 	tb.pos += sz
-}
-
-// PutField writes an interface{} to the ith field of the Tuple being built.
-func (tb *TupleBuilder) PutField(i int, v interface{}) {
-	if v == nil {
-		return // NULL
-	}
-
-	enc := tb.Desc.Types[i].Enc
-	switch enc {
-	case Int8Enc:
-		tb.PutInt8(i, int8(convInt(v)))
-	case Uint8Enc:
-		tb.PutUint8(i, uint8(convUint(v)))
-	case Int16Enc:
-		tb.PutInt16(i, int16(convInt(v)))
-	case Uint16Enc:
-		tb.PutUint16(i, uint16(convUint(v)))
-	case Int32Enc:
-		tb.PutInt32(i, int32(convInt(v)))
-	case Uint32Enc:
-		tb.PutUint32(i, uint32(convUint(v)))
-	case Int64Enc:
-		tb.PutInt64(i, int64(convInt(v)))
-	case Uint64Enc:
-		tb.PutUint64(i, uint64(convUint(v)))
-	case Float32Enc:
-		tb.PutFloat32(i, v.(float32))
-	case Float64Enc:
-		tb.PutFloat64(i, v.(float64))
-	case DecimalEnc:
-		tb.PutDecimal(i, v.(string))
-	case TimeEnc:
-		tb.PutSqlTime(i, v.(string))
-	case YearEnc:
-		tb.PutYear(i, v.(int16))
-	case DateEnc, DatetimeEnc, TimestampEnc:
-		tb.PutTimestamp(i, v.(time.Time))
-	case StringEnc:
-		tb.PutString(i, v.(string))
-	case BytesEnc:
-		if s, ok := v.(string); ok {
-			v = []byte(s)
-		}
-		tb.PutBytes(i, v.([]byte))
-	case GeometryEnc:
-		// todo(andy): remove GMS dependency
-		tb.PutGeometry(i, serializeGeometry(v))
-	case JSONEnc:
-		// todo(andy): remove GMS dependency
-		tb.PutJSON(i, v.(sql.JSONDocument).Val)
-	default:
-		panic(fmt.Sprintf("unknown encoding %v %v", enc, v))
-	}
-}
-
-func convInt(v interface{}) int {
-	switch i := v.(type) {
-	case int:
-		return i
-	case int8:
-		return int(i)
-	case uint8:
-		return int(i)
-	case int16:
-		return int(i)
-	case uint16:
-		return int(i)
-	case int32:
-		return int(i)
-	case uint32:
-		return int(i)
-	case int64:
-		return int(i)
-	case uint64:
-		return int(i)
-	default:
-		panic("impossible conversion")
-	}
-}
-
-func convUint(v interface{}) uint {
-	switch i := v.(type) {
-	case uint:
-		return i
-	case int:
-		return uint(i)
-	case int8:
-		return uint(i)
-	case uint8:
-		return uint(i)
-	case int16:
-		return uint(i)
-	case uint16:
-		return uint(i)
-	case int32:
-		return uint(i)
-	case uint32:
-		return uint(i)
-	case int64:
-		return uint(i)
-	case uint64:
-		return uint(i)
-	default:
-		panic("impossible conversion")
-	}
-}
-
-// todo(andy): remove GMS dependency
-//  have the engine pass serialized bytes
-
-const (
-	sridSize       = uint32Size
-	endianSize     = uint8Size
-	typeSize       = uint32Size
-	ewkbHeaderSize = sridSize + endianSize + typeSize
-)
-
-const (
-	pointType      = uint32(1)
-	linestringType = uint32(2)
-	polygonType    = uint32(3)
-
-	littleEndian = uint8(1)
-)
-
-type ewkbHeader struct {
-	srid   uint32
-	endian uint8
-	typ    uint32
-}
-
-func (h ewkbHeader) writeTo(buf []byte) {
-	expectSize(buf, ewkbHeaderSize)
-	writeUint32(buf[:sridSize], h.srid)
-	writeUint8(buf[sridSize:sridSize+endianSize], h.endian)
-	writeUint32(buf[sridSize+endianSize:ewkbHeaderSize], h.typ)
-}
-
-func readHeaderFrom(buf []byte) (h ewkbHeader) {
-	expectSize(buf, ewkbHeaderSize)
-	h.srid = readUint32(buf[:sridSize])
-	h.endian = readUint8(buf[sridSize : sridSize+endianSize])
-	h.typ = readUint32(buf[sridSize+endianSize : ewkbHeaderSize])
-	return
-}
-
-func serializeGeometry(v interface{}) []byte {
-	switch t := v.(type) {
-	case sql.Point:
-		return serializePoint(t)
-	case sql.Linestring:
-		return serializeLinestring(t)
-	case sql.Polygon:
-		return serializePolygon(t)
-	default:
-		panic(fmt.Sprintf("unknown geometry %v", v))
-	}
-}
-
-func serializePoint(p sql.Point) (buf []byte) {
-	pb := function.PointToBytes(p)
-	buf = make([]byte, ewkbHeaderSize+ByteSize(len(pb)))
-	copy(buf[ewkbHeaderSize:], pb)
-
-	h := ewkbHeader{
-		srid:   p.SRID,
-		endian: littleEndian,
-		typ:    pointType,
-	}
-	h.writeTo(buf[:ewkbHeaderSize])
-	return
-}
-
-func serializeLinestring(l sql.Linestring) (buf []byte) {
-	lb := function.LineToBytes(l)
-	buf = make([]byte, ewkbHeaderSize+ByteSize(len(lb)))
-	copy(buf[ewkbHeaderSize:], lb)
-
-	h := ewkbHeader{
-		srid:   l.SRID,
-		endian: littleEndian,
-		typ:    linestringType,
-	}
-	h.writeTo(buf[:ewkbHeaderSize])
-	return
-}
-
-func serializePolygon(p sql.Polygon) (buf []byte) {
-	pb := function.PolyToBytes(p)
-	buf = make([]byte, ewkbHeaderSize+ByteSize(len(pb)))
-	copy(buf[ewkbHeaderSize:], pb)
-
-	h := ewkbHeader{
-		srid:   p.SRID,
-		endian: littleEndian,
-		typ:    polygonType,
-	}
-	h.writeTo(buf[:ewkbHeaderSize])
-	return
 }
